Add levels of abstraction and named constants
A number of methods are too long and nested deeply.
Many numbers included in the current implementation are not explained.

This way of writing code makes the current code hard to understand.

Using helper method calls maintains the abstraction in different levels and
using named constants and referencing them in the code helps with readability.

Let's,
* add helper method calls where necessary to add layers of abstraction
* use named constants to avoid using magic numbers and improve understanding